Edited MemberEditSearch and adjusted the starting positions for forms

+++ Team5BLibraryManagementSystem/UC_Membersearchedit.cs
@@ -13,32 +13,46 @@ namespace Team5BLibraryManagementSystem
 {
     public partial class Uc_Membersearchedit : UserControl
     {
-        ZtkValidityCheck chkvldy = new ZtkValidityCheck();   
+        ZtkValidityCheck chkvldy = new ZtkValidityCheck();    // Class which contains methods for all the validity checkings
         public Uc_Membersearchedit()
         {
             InitializeComponent();
-            labelSearch.Visible = false; ;
+            labelSearch.Visible = false;
         }
 
+        public string memberId
+        {
+            get { return textbox_Idnumber.Text; }       
+        }
         private void btn_Print_Click(object sender, EventArgs e)
         {
 
         }
 
+
+        //Remove member
         private void btn_Delete_Click(object sender, EventArgs e)
         {
             SA47Team05BESNETLMSEntities context = new SA47Team05BESNETLMSEntities();
             var q = context.Members.Where(x => x.memberid.ToString() == textbox_Idnumber.Text).First();
-
-            DialogResult dialogResult = MessageBox.Show(String.Format("Are you sure you want to remove the user,{0}?", q.name), "Warning", MessageBoxButtons.YesNo);
-
-            if (dialogResult == DialogResult.Yes)
+            if(ZtkValidityCheck.MemberIsValid(q.expirydate))
             {
-                context.Members.Remove(q);
-                context.SaveChanges();
-                MessageBox.Show(String.Format("{0} has been removed.", q.name), "Confirmation", MessageBoxButtons.OK);
-                Clearfields();
+                MessageBox.Show("Deletion of active member is not allowed!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
+
+            else
+            {
+                DialogResult dialogResult = MessageBox.Show(String.Format("Are you sure you want to remove the user,{0}?", q.name), "Warning", MessageBoxButtons.YesNo);
+
+                if (dialogResult == DialogResult.Yes)
+                {
+                    context.Members.Remove(q);
+                    context.SaveChanges();
+                    MessageBox.Show(String.Format("{0} has been removed.", q.name), "Confirmation", MessageBoxButtons.OK);
+                    Clearfields();
+                }
+            }
+            
         }
 
         private void btn_Search_Click(object sender, EventArgs e)
@@ -59,6 +73,7 @@ namespace Team5BLibraryManagementSystem
                     switch (combobox_Category.SelectedItem)
                     {
                         case "Member ID":
+
                             if (ZtkValidityCheck.IsValidID(textbox_Category.Text))
                             {
                                 Gb_Memberinformation.Visible = true;
@@ -82,6 +97,7 @@ namespace Team5BLibraryManagementSystem
                                 }
                                 else
                                 {
+                                    btn_Renew.Visible = true;
                                     textbox_Status.Text="Expired";
                                 }
 
@@ -89,6 +105,7 @@ namespace Team5BLibraryManagementSystem
                             break;
 
                         case "Member Name":
+
                             if (ZtkValidityCheck.IsValidName(textbox_Category.Text))
                             {
 
@@ -113,6 +130,7 @@ namespace Team5BLibraryManagementSystem
                                 }
                                 else
                                 {
+                                    btn_Renew.Visible = true;
                                     textbox_Status.Text = "Expired";
                                 }
                             }
@@ -143,6 +161,7 @@ namespace Team5BLibraryManagementSystem
                                 }
                                 else
                                 {
+                                    btn_Renew.Visible = true;
                                     textbox_Status.Text = "Expired";
                                 }
                             }
@@ -172,6 +191,7 @@ namespace Team5BLibraryManagementSystem
                                 }
                                 else
                                 {
+                                    btn_Renew.Visible = true;
                                     textbox_Status.Text = "Expired";
                                 }
                             }
@@ -205,28 +225,89 @@ namespace Team5BLibraryManagementSystem
 
         private void btn_Edit_Click(object sender, EventArgs e)
         {
-            btn_Edit.Visible = false;
-            btn_Save.Visible = true;
-            btn_Cancel.Visible = true;
-            btn_Delete.Visible = false;
-            textbox_Address.Enabled = true;
-            textbox_Membername.Enabled = true;
-            textbox_Phone.Enabled = true;
-            textbox_email.Enabled = true;
-            datetimepicker_Dob.Enabled = true;
+            SA47Team05BESNETLMSEntities context = new SA47Team05BESNETLMSEntities();
+            var u = (from x in context.Members
+                     where x.memberid.ToString() == textbox_Idnumber.Text
+                     select x).First();
+
+            if (ZtkValidityCheck.MemberIsValid(u.expirydate))
+            {
+                btn_Edit.Visible = false;
+                btn_Save.Visible = true;
+                btn_Cancel.Visible = true;
+                btn_Delete.Visible = false;
+                textbox_Address.Enabled = true;
+                textbox_Membername.Enabled = true;
+                textbox_Phone.Enabled = true;
+                textbox_email.Enabled = true;
+                datetimepicker_Dob.Enabled = true;
+            }
+            else
+            {
+                DialogResult dialogresult = MessageBox.Show("Membership has been expired!\nEditing is not allowed!\nWould you like to renew the membership?", "Warning", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                if(dialogresult == DialogResult.Yes)
+                {
+                    RenewMembership();
+                }
+            }
+
+            
         }
 
         private void btn_Save_Click(object sender, EventArgs e)
         {
             try
             {
-                if (textbox_Membername.Text == "" || textbox_Phone.Text == "" || textbox_email.Text == "" || textbox_Address.Text == "")
+              
+                if(textbox_Membername.Text == "" || textbox_Phone.Text == "" || textbox_email.Text == "" || textbox_Address.Text == "")
                 {
-                    MessageBox.Show("Please fill up all the fields!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    if (textbox_Membername.Text == "")
+                    {
+                        label_MemberName.Visible = true;
+                        MessageBox.Show("Member name field cannot be left empty!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        label_MemberName.Visible = false;
+                    }
+
+
+                    if (textbox_Phone.Text == "")
+                    {
+                        label_Phone.Visible = true;
+                        MessageBox.Show("Phone number field cannot be left empty!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        label_Phone.Visible = false;
+                    }
+
+                    if (textbox_email.Text == "")
+                    {
+                        label_EmailAddress.Visible = true;
+                        MessageBox.Show("Email address field cannot be left empty!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        label_EmailAddress.Visible = false;
+                    }
+
+                    if (textbox_Address.Text == "")
+                    {
+                        label_Address.Visible = true;
+                        MessageBox.Show("Address field cannot be left empty!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        label_Address.Visible = false;
+                    }
                 }
+                
 
                 else
                 {
+                    WarningLabels(false);
+
                     SA47Team05BESNETLMSEntities context = new SA47Team05BESNETLMSEntities();
                     Member q = context.Members.Where(x => x.memberid.ToString() == textbox_Idnumber.Text).First();
 
@@ -261,6 +342,18 @@ namespace Team5BLibraryManagementSystem
             DefaultVisibility();
         }
 
+        //Set warning labels' Visibilities
+
+        private void WarningLabels(bool visibility)
+        {
+            label_EmailAddress.Visible = visibility;
+            label_Address.Visible = visibility;
+            label_MemberId.Visible = visibility;
+            label_MemberName.Visible = visibility;
+            label_Phone.Visible = visibility;
+        }
+
+        //Set Control's Visibility
         private void DefaultVisibility()
         {
             btn_Edit.Visible = true;
@@ -285,6 +378,44 @@ namespace Team5BLibraryManagementSystem
             datetimepicker_Dob.ResetText();
             datetimepicker_Expirydate.ResetText();
             datetimepicker_Joineddate.ResetText();
+            textbox_Status.Clear();
+        }
+
+        //Renew Membership Trigger
+        private void btn_Renew_Click(object sender, EventArgs e)
+        {
+            RenewMembership();
+        }
+
+        //Renew Membership
+        private void RenewMembership()
+        {
+            SA47Team05BESNETLMSEntities context = new SA47Team05BESNETLMSEntities();
+            var u = (from x in context.Members
+                     where x.memberid.ToString() == textbox_Idnumber.Text
+                     select x).First();
+
+            if (ZtkValidityCheck.MemberIsValid(u.expirydate))
+            {
+                MessageBox.Show("Membership is valid untill" + u.expirydate.Date + "!\nNo Need to renew!", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+
+            else
+            {
+                u.joindate = DateTime.Now;
+                u.expirydate = DateTime.Now.AddYears(1);
+                context.SaveChanges();
+                MessageBox.Show("Membership has been updated for " + u.name + "\nNew Expiry Date: " + u.expirydate + "\nThank you!", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            datetimepicker_Joineddate.Text = u.joindate.ToString();
+            datetimepicker_Expirydate.Text = u.expirydate.ToString();
+
+        }
+
+        private void btn_History_Click(object sender, EventArgs e)
+        {
+            Frm_TransactionHist transtForm = new Frm_TransactionHist(this);
+            transtForm.ShowDialog();
         }
     }
 }
